fix: lingui plural issue
