+++ web-marketplace/src/legacy-pages/ProjectReview/ProjectReview.tsx
@@ -206,6 +206,7 @@ export const ProjectReview: React.FC<React.PropsWithChildren<unknown>> = () => {
         <ItemDisplay name={_(msg`Size`)}>
           {metadata?.['regen:projectSize']?.['qudt:numericValue'] || '-'}{' '}
           {getAreaUnit(
+            _,
             metadata?.['regen:projectSize']?.['qudt:unit'] as qudtUnit,
             metadata?.['regen:projectSize']?.['qudt:numericValue'],
           )}
