+++ web-marketplace/src/lib/queries/react-query/registry-server/graphql/getProjectBySlugQuery/getProjectBySlugQuery.ts
@@ -57,6 +57,8 @@ export const getProjectBySlugQuery = ({
 
       return ProjectBySlug;
     } catch (e) {
+      console.log('e', e);
+
       return null;
     }
   },
